requirements.txt
Flask
flask-cors
python-dotenv
requests
gunicorn
pymysql
openai
pandas
opencv-python
numpy
Flask-JWT-Extended